프로필 화면 구현

See merge request swm-12/12_swm24/speakworld-android!52

+++ app/src/main/java/com/somasoma/wagglewaggle/presentation/auth/sign_in_and_sign_up/SignInAndSignUpViewModel.kt
@@ -100,7 +100,7 @@ class SignInAndSignUpViewModel @Inject constructor(
 
     private fun saveMemberId(firebaseResponse: FirebaseResponse) {
         firebaseResponse.memberId?.let {
-            sharedPreferenceHelper.putLong(
+            sharedPreferenceHelper.putInt(
                 PreferenceConstant.MEMBER_ID,
                 it
             )
